[refactor] message request용 dto 변수 수정 및 파일 재배치
