+++ src/main/java/com/gachigage/chat/WebSocketConfig.java
@@ -7,6 +7,8 @@ import org.springframework.web.socket.config.annotation.EnableWebSocketMessageBr
 import org.springframework.web.socket.config.annotation.StompEndpointRegistry;
 import org.springframework.web.socket.config.annotation.WebSocketMessageBrokerConfigurer;
 
+import com.gachigage.chat.service.StompHandler;
+
 import lombok.RequiredArgsConstructor;
 
 @Configuration
