+++ src/main/java/com/gachigage/chat/dto/ChatMessageRequestDto.java
@@ -15,7 +15,7 @@ public class ChatMessageRequestDto {
 
 	private String senderNickname;
 
-	private String message;
+	private String content;
 
 	private ChatMessageType messageType;
